Corregido tipo de datos a bigdecimal en la base de datos con los id , adicional corregido problema de autoincremental que se estaba incrementando de 100 en 100 y modificado orden de los iconos del escritorio

+++ workspace/servidor/src/main/java/ec/com/codesoft/codefaclite/servidor/entity/Impuesto.java
@@ -37,7 +37,7 @@ public class Impuesto implements Serializable
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "ID_IMPUESTO")
-    private Integer idImpuesto;
+    private Long idImpuesto;
     @Column(name = "NOMBRE")
     private String nombre;
     @Column(name = "CODIGO_SRI")
@@ -50,11 +50,11 @@ public class Impuesto implements Serializable
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "impuesto")
     private List<ImpuestoDetalle> detalleImpuestos;
 
-    public Integer getIdImpuesto() {
+    public Long getIdImpuesto() {
         return idImpuesto;
     }
 
-    public void setIdImpuesto(Integer idImpuesto) {
+    public void setIdImpuesto(Long idImpuesto) {
         this.idImpuesto = idImpuesto;
     }
